1. refractor when processing procesBetResult, if resultType = Win, Lose, and END, and operator is responding with insufficient_fund, we will always return success and resend the request to operator for them to retry process
2. idempotent bet_result handling whenever first bet_result is ending with insufficient_fund

+++ src/main/java/com/nextgen/gameaggregator/service/WalletService.java
@@ -300,6 +300,7 @@ public class WalletService {
 
         // send bet data to Operator
         try {
+            walletBetResultData.setBalance(settledBet.getBalance());
             balanceVo = walletBetResultAction.call(traceId, agentId, gameSession, walletBetResultData, resultType, httpRequestLog, fromVendorConversionRate, toVendorConversionRate);
 
             loggingService.logStart();
@@ -529,6 +530,7 @@ public class WalletService {
 
                 try {
                     // record operator processing time
+                    walletBetResultData.setBalance(unsettledBet.getBalance());
                     balanceVo = walletBetResultAction.call(traceId, agentId, gameSession, walletBetResultData, resultType, httpRequestLog, fromVendorConversionRate, toVendorConversionRate);
                     BigDecimal balance = balanceVo.getData().getBalance();
 
@@ -549,9 +551,11 @@ public class WalletService {
 
                     loggingService.logStart();
 
-                    if (operatorStatus == ResponseCodes.Status.SC_INSUFFICIENT_FUNDS.code) {
-                        unsettledBetService.deleteWithoutClearingCache(unsettledBet);
-
+                    if (resultType.equals(ResultType.BET_WIN) || resultType.equals(ResultType.BET_LOSE)) {
+                        if (operatorStatus == ResponseCodes.Status.SC_INSUFFICIENT_FUNDS.code) {
+                            cachingService.updateUnsettledBetCaching(unsettledBet);
+                            unsettledBetService.deleteWithoutClearingCache(unsettledBet);
+                        }
                     } else {
                         betResultLogService.save(rawBetResultLog);
                         unsettledBetService.save(unsettledBet);
